dash_ps: separate to util, wallet, net submodules
