+++ electrum_dash/gui/kivy/uix/dialogs/coins_dialog.py
@@ -7,8 +7,8 @@ from kivy.uix.recycleview.views import RecycleDataViewBehavior
 from kivy.uix.recycleboxlayout import RecycleBoxLayout
 from kivy.uix.behaviors import FocusBehavior
 
-from electrum_dash.dash_ps import (PSCoinRounds, ps_coin_rounds_str,
-                                   sort_utxos_by_ps_rounds)
+from electrum_dash.dash_ps_util import (PSCoinRounds, ps_coin_rounds_str,
+                                        sort_utxos_by_ps_rounds)
 from electrum_dash.gui.kivy.i18n import _
 from electrum_dash.gui.kivy.uix.context_menu import ContextMenu
